Things

+++ LOTR/src/main/java/lotr/client/gui/LOTRGuiScreenBase.java
@@ -51,10 +51,10 @@ public abstract class LOTRGuiScreenBase extends GuiScreen
         float f1 = 1F / (float)imageWidth;
         Tessellator tessellator = Tessellator.instance;
         tessellator.startDrawingQuads();
-        tessellator.addVertexWithUV(x + 0, y + height, zLevel, (float)(u + 0) * f, (float)(v + height) * f1);
-        tessellator.addVertexWithUV(x + width, y + height, zLevel, (float)(u + width) * f, (float)(v + height) * f1);
-        tessellator.addVertexWithUV(x + width, y + 0, zLevel, (float)(u + width) * f, (float)(v + 0) * f1);
-        tessellator.addVertexWithUV(x + 0, y + 0, zLevel, (float)(u + 0) * f, (float)(v + 0) * f1);
+        tessellator.addVertexWithUV(x + 0, y + height, zLevel, (u + 0) * f, (v + height) * f1);
+        tessellator.addVertexWithUV(x + width, y + height, zLevel, (u + width) * f, (v + height) * f1);
+        tessellator.addVertexWithUV(x + width, y + 0, zLevel, (u + width) * f, (v + 0) * f1);
+        tessellator.addVertexWithUV(x + 0, y + 0, zLevel, (u + 0) * f, (v + 0) * f1);
         tessellator.draw();
     }
 }
